requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
pyautogen==0.2.0
google-generativeai==0.3.2
arxiv==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
jupyter==1.0.0
chromadb==0.5.5
langchain-text-splitters==0.3.1
pypdf==4.2.0
faiss-cpu==1.8.0.post1
